Custom text metrics provider
